+++ src/PockyBot.NET/Services/DirectResultsMessageSender.cs
@@ -1,7 +1,9 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using GlobalX.ChatBots.Core.Messages;
+using Microsoft.Extensions.Logging;
 using PockyBot.NET.Models;
 
 namespace PockyBot.NET.Services
@@ -9,17 +11,26 @@ namespace PockyBot.NET.Services
     internal class DirectResultsMessageSender : IDirectResultsMessageSender
     {
         private readonly IMessageHandler _messageHandler;
+        private readonly ILogger<DirectResultsMessageSender> _logger;
 
-        public DirectResultsMessageSender(IMessageHandler messageHandler)
+        public DirectResultsMessageSender(IMessageHandler messageHandler, ILogger<DirectResultsMessageSender> logger)
         {
             _messageHandler = messageHandler;
+            _logger = logger;
         }
 
         public async Task SendDirectMessagesAsync(List<PegRecipient> recipients)
         {
             foreach (var recipient in recipients)
             {
-                await SendDirectMessageAsync(recipient).ConfigureAwait(false);
+                try
+                {
+                    await SendDirectMessageAsync(recipient).ConfigureAwait(false);
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError($"Error sending message to {recipient.Name}", ex);
+                }
             }
         }
 
@@ -43,7 +54,7 @@ namespace PockyBot.NET.Services
         {
             var location = peg.SenderLocation ?? "No Location";
             var pointsPlural = peg.Weight == 1 ? string.Empty : "s";
-            return $"* **{peg.SenderName}** ({location}, {peg.Weight} point{pointsPlural}) — \"_{peg.Comment}_\"";
+            return $"* **{peg.SenderName}** ({location}, {peg.Weight} point{pointsPlural}) — \"_{peg.Comment.Replace("\n\n", " // ").Replace("\n", " / ")}_\"";
         }
     }
 }