+++ src/PockyBot.NET/Services/Pegs/PegResultsHelper.cs
@@ -30,7 +30,7 @@ namespace PockyBot.NET.Services.Pegs
                 var receiverLocation = x.Location?.Location;
                 var validPegs = GetValidPegs(x.PegsReceived, requireKeywords, keywords, penaltyKeywords,
                     linkedKeywords, receiverLocation);
-                var penaltyPegs = GetPenaltyPegs(x.PegsGiven, requireKeywords, keywords.Concat(linkedKeywords.Select(x =>  x.Keyword)).ToArray(), penaltyKeywords);
+                var penaltyPegs = GetPenaltyPegs(x.PegsGiven, keywords.Concat(linkedKeywords.Select(k =>  k.Keyword)).ToArray(), penaltyKeywords);
 
                 return new PegRecipient
                 {
@@ -105,10 +105,9 @@ namespace PockyBot.NET.Services.Pegs
                 .ToList();
         }
 
-        private List<PegDetails> GetPenaltyPegs(List<Peg> pegs, int? requireKeywords, string[] keywords,
-            string[] penaltyKeywords)
+        private List<PegDetails> GetPenaltyPegs(List<Peg> pegs, string[] keywords, string[] penaltyKeywords)
         {
-            return pegs.Where(y => _pegHelper.IsPenaltyPeg(y.Comment, requireKeywords, penaltyKeywords, keywords))
+            return pegs.Where(y => _pegHelper.IsPenaltyPeg(y.Comment, penaltyKeywords, keywords))
                 .Select(y => new PegDetails
                 {
                     SenderName = y.Receiver.Username,