+++ src/PockyBot.NET/Services/Triggers/Peg.cs
@@ -46,6 +46,76 @@ namespace PockyBot.NET.Services.Triggers
 
         public async Task<Message> Respond(Message message)
         {
+            if (!IsPegFormatValid(message, out var pegFormatErrorMessage))
+            {
+                return pegFormatErrorMessage;
+            }
+
+            _logger.LogDebug("Getting user information for id {userid}, username {username}", message.Sender.UserId,
+                message.Sender.Username);
+            var sender = await _pockyUserRepository.AddOrUpdateUserAsync(message.Sender.UserId, message.Sender.Username);
+
+            var penaltyKeywords = _configRepository.GetStringConfig("penaltyKeyword").ToArray();
+            var keywords = GetKeywords();
+
+            var comment = string.Join(string.Empty, message.MessageParts.Skip(3).Select(x => x.Text)).Trim();
+            var isPenaltyPeg = _pegHelper.IsPenaltyPeg(comment, penaltyKeywords, keywords);
+            var numPegsGiven = sender.PegsGiven?
+                .Count(x => !_pegHelper.IsPenaltyPeg(x.Comment, penaltyKeywords, keywords)) ?? 0;
+
+
+            if (!UserHasPegsRemaining(sender, isPenaltyPeg, numPegsGiven, out var pegsRemainingErrorMessage))
+            {
+                return pegsRemainingErrorMessage;
+            }
+
+            if (!IsPegMessageValid(message, out var pegMessageErrorMessage))
+            {
+                return pegMessageErrorMessage;
+            }
+
+            await GivePeg(message, comment, sender, isPenaltyPeg, numPegsGiven);
+            return null;
+        }
+
+        private bool IsPegFormatValid(Message message, out Message errorMessage)
+        {
+            try
+            {
+                _logger.LogDebug("Validating peg request format");
+                _pegRequestValidator.ValidatePegRequestFormat(message);
+            }
+            catch (PegValidationException e)
+            {
+                _logger.LogWarning("Peg message format is invalid: {@error}", e);
+                errorMessage = new Message
+                {
+                    Text = e.Message
+                };
+                return false;
+            }
+            errorMessage = null;
+            return true;
+        }
+
+        private bool UserHasPegsRemaining(PockyUser sender, bool isPenaltyPeg, int numPegsGiven, out Message errorMessage)
+        {
+            if (!sender.HasRole(Role.Unmetered) && !isPenaltyPeg &&
+                (numPegsGiven >= _configRepository.GetGeneralConfig("limit")))
+            {
+                _logger.LogDebug("User {userId} has reached their peg limit", sender.UserId);
+                errorMessage = new Message
+                {
+                    Text = "Sorry, but you have already spent all of your pegs for this fortnight."
+                };
+                return false;
+            }
+            errorMessage = null;
+            return true;
+        }
+
+        private bool IsPegMessageValid(Message message, out Message errorMessage)
+        {
             try
             {
                 _logger.LogDebug("Validating peg message");
@@ -54,46 +124,33 @@ namespace PockyBot.NET.Services.Triggers
             catch (PegValidationException e)
             {
                 _logger.LogWarning("Peg message is invalid: {@error}", e);
-                return new Message
+                errorMessage = new Message
                 {
                     Text = e.Message
                 };
+                return false;
             }
+            errorMessage = null;
+            return true;
+        }
 
-            _logger.LogDebug("Getting user information for id {userid}, username {username}", message.Sender.UserId,
-                message.Sender.Username);
-            var sender = await _pockyUserRepository.AddOrUpdateUserAsync(message.Sender.UserId, message.Sender.Username);
-            var comment = string.Join(string.Empty, message.MessageParts.Skip(3).Select(x => x.Text)).Trim();
-
-            var requireKeywords = _configRepository.GetGeneralConfig("requireValues");
+        private string[] GetKeywords()
+        {
             var keywords = _configRepository.GetStringConfig("keyword").ToArray();
-            var penaltyKeywords = _configRepository.GetStringConfig("penaltyKeyword").ToArray();
             var linkedKeywords = _configRepository.GetStringConfig("linkedKeyword").Select(x => new LinkedKeyword(x));
-            keywords = keywords.Concat(linkedKeywords.Select(x => x.LinkedWord)).ToArray();
+            return keywords.Concat(linkedKeywords.Select(x => x.LinkedWord)).ToArray();
+        }
 
-            var isPegValid = _pegHelper.IsPegValid(comment, requireKeywords, keywords, penaltyKeywords);
-            var numPegsGiven = sender.PegsGiven?.Count(x =>
-                                   _pegHelper.IsPegValid(x.Comment, requireKeywords, keywords, penaltyKeywords)) ?? 0;
-
-            if (!sender.HasRole(Role.Unmetered) &&
-                (isPegValid && numPegsGiven >= _configRepository.GetGeneralConfig("limit")))
-            {
-                _logger.LogDebug("User {userId} has reached their peg limit", sender.UserId);
-                return new Message
-                {
-                    Text = "Sorry, but you have already spent all of your pegs for this fortnight."
-                };
-            }
-
+        private async Task GivePeg(Message message, string comment, PockyUser sender, bool isPenaltyPeg, int numPegsGiven)
+        {
             var receiverId = message.MessageParts[2].UserId;
             var receiver = await _chatHelper.People.GetPersonAsync(receiverId).ConfigureAwait(false);
             var dbReceiver = await _pockyUserRepository.AddOrUpdateUserAsync(receiver.UserId, receiver.Username);
 
             _logger.LogInformation(
-                "Giving peg with sender {senderId}, receiver {receiverId}, validity {isPegValid}, comment {comment}",
-                sender.UserId, receiver.UserId, isPegValid, comment);
-            await _pegGiver.GivePeg(comment, sender, dbReceiver, isPegValid ? numPegsGiven + 1 : numPegsGiven).ConfigureAwait(false);
-            return null;
+                "Giving peg with sender {senderId}, receiver {receiverId}, is penalty peg? {isPegValid}, comment {comment}",
+                sender.UserId, receiver.UserId, isPenaltyPeg, comment);
+            await _pegGiver.GivePeg(comment, sender, dbReceiver, !isPenaltyPeg ? numPegsGiven + 1 : numPegsGiven).ConfigureAwait(false);
         }
     }
 }