+++ src/PockyBot.NET/Services/UsernameUpdater.cs
@@ -1,7 +1,9 @@
 using System.Collections.Generic;
 using System.Linq;
+using System.Net.Http;
 using System.Threading.Tasks;
 using GlobalX.ChatBots.Core.People;
+using Microsoft.Extensions.Logging;
 using PockyBot.NET.Persistence.Models;
 using PockyBot.NET.Persistence.Repositories;
 
@@ -11,32 +13,45 @@ namespace PockyBot.NET.Services
     {
         private readonly IPersonHandler _personHandler;
         private readonly IPockyUserRepository _pockyUserRepository;
+        private readonly ILogger<UsernameUpdater> _logger;
 
-        public UsernameUpdater(IPersonHandler personHandler, IPockyUserRepository pockyUserRepository)
+        public UsernameUpdater(IPersonHandler personHandler, IPockyUserRepository pockyUserRepository, ILogger<UsernameUpdater> logger)
         {
             _personHandler = personHandler;
             _pockyUserRepository = pockyUserRepository;
+            _logger = logger;
         }
 
         public async Task<List<PockyUser>> UpdateUsernames(List<PockyUser> users)
         {
-            var usersDetails = await Task.WhenAll(users.Select(x => _personHandler.GetPersonAsync(x.UserId)));
-            var dbUpdates = new List<Task>();
+            var usersDetails = await Task.WhenAll(users.Select(x => GetUser(x.UserId, x.Username)));
 
-            var updatedUsers = users.Select(u =>
+            var updatedUsers = new List<PockyUser>();
+            foreach (var user in users)
             {
-                var userDetails = usersDetails.FirstOrDefault(x => x.UserId == u.UserId);
-                if (userDetails?.Username != u.Username)
+                var userDetails = usersDetails.FirstOrDefault(x => x.UserId == user.UserId);
+                if (userDetails?.Username != user.Username)
                 {
-                    u.Username = userDetails?.Username;
-                    dbUpdates.Add(_pockyUserRepository.UpdateUsernameAsync(u.UserId, u.Username));
+                    user.Username = userDetails?.Username;
+                    await _pockyUserRepository.UpdateUsernameAsync(user.UserId, user.Username);
                 }
+                updatedUsers.Add(user);
+            }
 
-                return u;
-            }).ToList();
-
-            await Task.WhenAll(dbUpdates).ConfigureAwait(false);
             return updatedUsers;
+        }
+
+        private async Task<Person> GetUser(string userId, string username)
+        {
+            try
+            {
+                return await _personHandler.GetPersonAsync(userId);
+            }
+            catch (HttpRequestException ex)
+            {
+                _logger.LogError($"Error retrieving details for user {userId}", ex);
+                return new Person{UserId = userId, Username = username};
+            }
         }
     }
 }